requirements.txt
ultralytics>=8.0.0
pillow>=9.0.0
opencv-python>=4.5.0
pyyaml>=5.4.0
numpy>=1.19.0
pytest>=7.0.0
python-dotenv>=0.19.0
jupyter>=1.0.0
